requirements.txt
cryptography==42.0.5
python-dotenv==1.0.1
requests==2.32.3
